+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
+
+## [3.x]
+### Added
+- Added `AbstractCategoriesBlockService` class
+- Added `AbstractCollectionsBlockService` class
+- Added `AbstractTagsBlockService` class